Create File Input component
